Rename contract_class module
