+++ starknet_py/hash/class_hash_test.py
@@ -3,7 +3,7 @@
 import pytest
 
 from starknet_py.common import create_contract_class
-from starknet_py.hash.contract_class import compute_class_hash
+from starknet_py.hash.class_hash import compute_class_hash
 from starknet_py.tests.e2e.fixtures.misc import read_contract
